requirements.txt
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.20.0
openai>=1.0.0
